+++ src/movies/MovieCrew.cpp
@@ -57,7 +57,7 @@ void MovieCrew::addActor(Actor actor)
 
 void MovieCrew::removeActor(Actor* actor)
 {
-    for (int i = 0, n = m_actors.size(); i < n; ++i) {
+    for (size_t i = 0, n = m_actors.size(); i < n; ++i) {
         if (m_actors[i].get() == actor) {
             m_actors.erase(m_actors.begin() + i);
             break;
